requirements.txt
Django==5.1.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-cors-headers>=4.3.1
drf-yasg==1.21.7
django-celery-beat>=2.5.0
redis>=5.0.1
celery>=5.3.6
python-dateutil==2.8.2
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.3.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
django-stubs>=4.2.0
python-dotenv>=1.0.0
Pillow==10.2.0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
django-ratelimit==4.1.0
django-environ==0.11.2
prometheus-client==0.19.0
django-prometheus==2.3.1
python-json-logger==2.0.7
structlog==24.1.0
django-structlog==4.0.0
django-health-check==3.17.0 